Move base runner functionality into a new abstract class
