+++ Source/EventFlow.EventStores.EventStore.Tests/EventFlow.EventStores.EventStore.Tests.csproj
@@ -45,7 +45,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="EventStoreRunner.cs" />
-    <Compile Include="Extensions\ProcessExtensions.cs" />
     <Compile Include="IntegrationTests\EventStoreEventStoreTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="TcpHelper.cs" />
@@ -302,5 +301,8 @@
       <Private>True</Private>
       <Paket>True</Paket>
     </Reference>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Extensions\" />
   </ItemGroup>
 </Project>
