+++ Source/EventFlow.EventStores.EventStore.Tests/EventStoreRunner.cs
@@ -24,15 +24,11 @@
 
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
-using System.IO.Compression;
 using System.Linq;
 using System.Net;
-using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using EventFlow.EventStores.EventStore.Tests.Extensions;
 using EventFlow.Extensions;
 using EventFlow.TestHelpers;
 using EventStore.ClientAPI;
@@ -42,13 +38,15 @@ using NUnit.Framework;
 namespace EventFlow.EventStores.EventStore.Tests
 {
     [Category(Categories.Integration)]
-    public class EventStoreRunner
+    public class EventStoreRunner : Runner
     {
-        private static readonly Dictionary<Version, Uri> EventStoreVersions = new Dictionary<Version, Uri>
-        {
-            {new Version(3, 4, 0), new Uri("http://download.geteventstore.com/binaries/EventStore-OSS-Win-v3.4.0.zip", UriKind.Absolute)},
-            {new Version(3, 3, 1), new Uri("http://download.geteventstore.com/binaries/EventStore-OSS-Win-v3.3.1.zip", UriKind.Absolute)}
-        };
+        protected override IEnumerable<SoftwareDescription> SoftwareDescriptions { get; } = new[]
+            {
+                new SoftwareDescription(new Version(3, 3, 1), new Uri("http://download.geteventstore.com/binaries/EventStore-OSS-Win-v3.3.1.zip", UriKind.Absolute)),
+                new SoftwareDescription(new Version(3, 4, 0), new Uri("http://download.geteventstore.com/binaries/EventStore-OSS-Win-v3.4.0.zip", UriKind.Absolute)),
+            };
+
+        protected override string SoftwareName { get; } = "EventStore";
 
         public class EventStoreInstance : IDisposable
         {
@@ -80,10 +78,15 @@ namespace EventFlow.EventStores.EventStore.Tests
             }
         }
 
-        public static async Task<EventStoreInstance> StartAsync()
+        public static Task<EventStoreInstance> StartAsync()
         {
-            var eventStoreVersion = EventStoreVersions.OrderByDescending(kv => kv.Key).First();
-            await InstallEventStoreAsync(eventStoreVersion.Key).ConfigureAwait(false);
+            return new EventStoreRunner().InternalStartAsync();
+        }
+
+        private async Task<EventStoreInstance> InternalStartAsync()
+        {
+            var eventStoreVersion = SoftwareDescriptions.OrderByDescending(kv => kv.Version).First();
+            await InstallEventStoreAsync(eventStoreVersion.Version).ConfigureAwait(false);
 
             var tcpPort = TcpHelper.GetFreePort();
             var httpPort = TcpHelper.GetFreePort();
@@ -93,7 +96,7 @@ namespace EventFlow.EventStores.EventStore.Tests
             try
             {
                 processDisposable = StartExe(
-                    Path.Combine(GetEventStorePath(eventStoreVersion.Key), "EventStore.ClusterNode.exe"),
+                    Path.Combine(GetEventStorePath(eventStoreVersion.Version), "EventStore.ClusterNode.exe"),
                     "'admin' user added to $users",
                     "--mem-db=True",
                     "--cluster-size=1",
@@ -136,156 +139,6 @@ namespace EventFlow.EventStores.EventStore.Tests
             return new EventStoreInstance(
                 connectionStringUri,
                 processDisposable);
-        }
-
-        private static IDisposable StartExe(
-            string exePath,
-            string initializationDone,
-            params string[] arguments)
-        {
-            var process = new Process
-            {
-                StartInfo = new ProcessStartInfo(exePath, string.Join(" ", arguments))
-                {
-                    CreateNoWindow = true,
-                    UseShellExecute = false,
-                    RedirectStandardError = true,
-                    RedirectStandardOutput = true,
-                    WorkingDirectory = Path.GetDirectoryName(exePath),
-                }
-            };
-            var exeName = Path.GetFileName(exePath);
-            process.OutputDataReceived += (sender, eventArgs) =>
-            {
-                if (string.IsNullOrEmpty(eventArgs.Data))
-                {
-                    return;
-                }
-                Console.WriteLine("OUT - {0}: {1}", exeName, eventArgs.Data);
-            };
-            process.ErrorDataReceived += (sender, eventArgs) =>
-            {
-                if (string.IsNullOrEmpty(eventArgs.Data))
-                {
-                    return;
-                }
-                Console.WriteLine("ERR - {0}: {1}", exeName, eventArgs.Data);
-            };
-            Action<Process> initializeProcess = p =>
-                {
-                    Console.WriteLine($"{exeName} START =======================================");
-                    p.Start();
-                    p.BeginOutputReadLine();
-                    p.BeginErrorReadLine();
-                };
-            process.WaitForOutput(initializationDone, initializeProcess);
-
-            return new DisposableAction(() =>
-                {
-                    try
-                    {
-                        process.Kill();
-                        process.WaitForExit(10000);
-                        Console.WriteLine($"{process.ProcessName} KILLED  =======================================");
-                    }
-                    catch (Exception e)
-                    {
-                        Console.WriteLine($"Failed to kill process: {e.Message}");
-                    }
-                });
-        }
-
-        private static async Task InstallEventStoreAsync(Version version)
-        {
-            if (IsEventStoreInstalled(version))
-            {
-                Console.WriteLine($"EventStore v'{version}' is already installed");
-                return;
-            }
-
-            Console.WriteLine($"EventStore v{version} not installed, installing it");
-
-            var tempDownload = Path.Combine(
-                Path.GetTempPath(),
-                $"eventstore-v{version}-{Guid.NewGuid().ToString("N")}.zip");
-            try
-            {
-                await DownloadFileAsync(EventStoreVersions[version], tempDownload).ConfigureAwait(false);
-                var eventStorePath = GetEventStorePath(version);
-                ExtractZipFile(tempDownload, eventStorePath);
-            }
-            finally
-            {
-                if (File.Exists(tempDownload))
-                {
-                    File.Delete(tempDownload);
-                }
-            }
-        }
-
-        private static void ExtractZipFile(string zipSourcePath, string directoryDestinationPath)
-        {
-            Console.WriteLine($"Extracting '{zipSourcePath}' to '{directoryDestinationPath}'");
-
-            if (!Directory.Exists(directoryDestinationPath))
-            {
-                Directory.CreateDirectory(directoryDestinationPath);
-            }
-
-            ZipFile.ExtractToDirectory(zipSourcePath, directoryDestinationPath);
-        }
-
-        private static bool IsEventStoreInstalled(Version version)
-        {
-            return Directory.Exists(GetEventStorePath(version));
-        }
-
-        private static string GetEventStorePath(Version version)
-        {
-            return Path.Combine(
-                Path.GetTempPath(),
-                $"eventflow-eventstore-v{version}");
-        }
-
-        private static async Task DownloadFileAsync(Uri sourceUri, string destinationPath)
-        {
-            if (File.Exists(destinationPath))
-            {
-                throw new ArgumentException($"File '{destinationPath}' already exists");
-            }
-
-            Console.WriteLine($"Downloading '{sourceUri}' to '{destinationPath}'");
-
-            using (var httpClient = new HttpClient())
-            using (var httpRequestMessage = new HttpRequestMessage(HttpMethod.Get, sourceUri))
-            using (var httpResponseMessage = await httpClient.SendAsync(httpRequestMessage).ConfigureAwait(false))
-            {
-                if (!httpResponseMessage.IsSuccessStatusCode)
-                {
-                    throw new HttpRequestException($"Failed to download '{sourceUri}' due to '{httpResponseMessage.StatusCode}'");
-                }
-
-                using (var sourceStream = await httpResponseMessage.Content.ReadAsStreamAsync().ConfigureAwait(false))
-                using (var destinationStream = new FileStream(destinationPath, FileMode.CreateNew))
-                {
-                    await sourceStream.CopyToAsync(destinationStream).ConfigureAwait(false);
-                }
-            }
-        }
-
-        private class DisposableAction : IDisposable
-        {
-            private readonly Action _action;
-
-            public DisposableAction(Action action)
-            {
-                _action = action;
-            }
-
-            public void Dispose()
-            {
-                _action();
-            }
         }
     }
 }
